Fix: Complete rewrite of NormalizePath
The former algorithm was based on the psudocode from the RFC.
The new alogorithm uses two flags:
* isAbsolute - recognize that if the path starts out
as an absolute path, it needs to stay that way.
* atDirectoryLevel - recognize that if we encounter
a "." or ".." then it well be reduced by simply discarding it or going back /up one step
but then we will be in a directory context meaning that should we end the path at this point there needs
to be an empty-string segment to mark that the end of the path is reaching into a directory, not just
referring to the directory.